services.cms/app/requirements.txt
Django==4.1.2
# psycopg2-binary==2.9.3
dnspython==2.2.1
djangorestframework==3.13.1
debugpy==1.6.0
django-cors-headers==3.11.0 # CORS
django-rest-authemail==2.1.2
pymysql==1.0.2
mysqlclient==2.1.0
pytz==2022.1
djangorestframework_simplejwt==5.2.0
boto3==1.24.12
django-environ==0.9.0
django-apscheduler==0.6.2
fcm-django==1.0.12
Pillow==9.2.0
# pytesseract==0.3.10
googlemaps==4.6.0
DateTimeRange==1.2.0